+++ api/urls.py
@@ -8,6 +8,7 @@ from django.conf.urls import url, include
 from rest_framework import routers
 
 from api.views import test_view_with_arg, TestReverseViewSet
+from seed.data_importer.views import ImportFileViewSet
 from seed.data_importer.views import (
     handle_s3_upload_complete,
     get_upload_details,
@@ -17,13 +18,12 @@ from seed.data_importer.views import (
 from seed.views.api import get_api_schema
 from seed.views.cycles import CycleView
 from seed.views.datasets import DatasetViewSet
-from seed.views.import_files import ImportFileViewSet
+from seed.views.labels import LabelViewSet, UpdateInventoryLabelsAPIView
 from seed.views.main import DataFileViewSet, version, progress
 from seed.views.organizations import OrganizationViewSet
 from seed.views.projects import ProjectViewSet
-from seed.views.users import UserViewSet
 from seed.views.properties import PropertyViewSet, TaxLotViewSet
-from seed.views.labels import LabelViewSet, UpdateInventoryLabelsAPIView
+from seed.views.users import UserViewSet
 
 api_v2_router = routers.DefaultRouter()
 api_v2_router.register(r'datasets', DatasetViewSet, base_name="datasets")
